+++ build.gradle.kts
@@ -3,16 +3,19 @@ import com.google.protobuf.gradle.id
 group = "org.example"
 version = "1.0-SNAPSHOT"
 
-val armeriaVersion by extra("1.22.1")
+val armeriaVersion by extra("1.23.1")
+val googleProtobufVersion by extra("3.23.2")
 val grpcVersion by extra("1.55.1")
 val grpcKotlinVersion by extra("1.3.0")
-val protobufVersion by extra("3.23.2")
 val slf4jVersion by extra("1.7.36")
 
 plugins {
-    kotlin("jvm") version "1.8.21"
+    val kotlinVersion = "1.8.21"
+    val protobufVersion = "0.9.3"
+
+    kotlin("jvm") version kotlinVersion
     application
-    id("com.google.protobuf") version "0.9.3"
+    id("com.google.protobuf") version protobufVersion
 }
 
 apply(plugin = "com.google.protobuf")
@@ -23,7 +26,7 @@ repositories {
 
 protobuf {
     protoc {
-        artifact = "com.google.protobuf:protoc:$protobufVersion"
+        artifact = "com.google.protobuf:protoc:$googleProtobufVersion"
     }
 
     plugins {
@@ -55,16 +58,20 @@ protobuf {
 }
 
 dependencies {
+    // armeria
     implementation(platform("com.linecorp.armeria:armeria-bom:$armeriaVersion"))
     implementation("com.linecorp.armeria:armeria-grpc:$armeriaVersion")
     implementation("com.linecorp.armeria:armeria:$armeriaVersion}")
 
+    // protobuf
     implementation("io.grpc:grpc-kotlin-stub:$grpcKotlinVersion")
     implementation("io.grpc:grpc-protobuf:$grpcVersion")
-    implementation("com.google.protobuf:protobuf-kotlin:$protobufVersion")
+    implementation("com.google.protobuf:protobuf-kotlin:$googleProtobufVersion")
 
+    // slf4j
     runtimeOnly("org.slf4j:slf4j-simple:$slf4jVersion")
 
+    // test
     testImplementation(kotlin("test"))
     testImplementation("com.linecorp.armeria:armeria-junit5:$armeriaVersion")
 }
@@ -78,5 +85,5 @@ kotlin {
 }
 
 application {
-    mainClass.set("com.example.grpc.MainKt")
+    mainClass.set("com.example.grpc.Application")
 }
